// frontend/src/types/pharma.ts
export enum UnitOfMeasure {
  KG = 'kg',
  G = 'g',
  MG = 'mg',
  L = 'L',
  ML = 'ml',
  UNITS = 'units'
}

/** Estado de almacenamiento */
export enum StorageCondition {
  AMBIENT = 'ambient',         // 15-25°C
  REFRIGERATED = 'refrigerated', // 2-8°C
  FROZEN = 'frozen',           // -20°C o menos
  CONTROLLED = 'controlled'    // Temperatura controlada específica
}